tests/stripe/requirements-test.txt
pytest>=8.0
pytest-xdist>=3.5
pytest-md-report>=0.6
httpx>=0.27
stripe>=11.0
tenacity>=9.0
pyjwt[crypto]>=2.9
cryptography>=43.0
playwright>=1.48
